+++ docs/include/samplecode/SpotifyDiscovery/DiscoverDevices.py
@@ -5,13 +5,12 @@ try:
     print("Test Starting\n")
 
     # create a new instance of the discovery class.
-    # we will verify device connections, as well as print device details
-    # to the console as they are discovered.
-    discovery:SpotifyDiscovery = SpotifyDiscovery(True, printToConsole=True)
+    # we will print device details to the console as they are discovered.
+    discovery:SpotifyDiscovery = SpotifyDiscovery(printToConsole=True)
 
     # discover Spotify Connect devices on the network, waiting up to 
-    # 5 seconds for all devices to be discovered.
-    discovery.DiscoverDevices(timeout=5)
+    # 2 seconds for all devices to be discovered.
+    discovery.DiscoverDevices(timeout=2)
             
     # print all discovered devices.
     print("\n%s" % (discovery.ToString(True)))